// src/client/types.gen.ts
export type LeadFormApi = {
    LeadFormId?: number;
    FormName?: string | null;
    CaseTypeId?: number;
    CaseTypeName?: string | null;
    Enabled?: boolean;
    CreatedDate?: string;
    CreatedById?: number;
    CreatedBy?: string | null;
    DaysToExpiration?: number;
};
// ...
export type LeadRoleApi = {
    LeadRoleId?: number;
    RoleName?: string | null;
    IsOwner?: boolean;
    IsDefault?: boolean;
    PriorityOrder?: number;
};

export type LeadRoleStatusApi = {
    LeadRoleId?: number;
    VisibilityOnLead?: number;
};
// ...
export type LookupTypes = 'LeadSource' | 'CaseType' | 'MarketingSource' | 'Statuses' | 'Offices' | 'Forms' | 'Tags' | 'PhoneNumbers';
// ...
export type ReferredSubstatusType = 'ReferredOther' | 'ReferredDeclined' | 'ReferredPending' | 'ReferredAccepted' | 'ReferredClosed' | 'ReferredSigned' | 'ReferredNoFollowUp' | 'ReferredShowToReferralPartners';
// ...
export type SettingsOptions = {
    Name?: string | null;
    IsEnabled?: boolean;
};
// ...
export type StatusApi = {
    Id?: number;
    Status?: string | null;
    StatusName?: string | null;
    StatusScript?: string | null;
    IsClosed?: boolean;
    IsCurrentClient?: boolean;
    IsReferrable?: boolean;
    IsSequenced?: boolean;
    IsScheduleable?: boolean;
    IsIncludedInSearch?: boolean;
    IsValueEnabled?: boolean;
    IsOnDashboard?: boolean;
    IsResolved?: boolean;
    IsWanted?: boolean;
    DisplayOrder?: number;
    Substatuses?: Array<SubStatusApi> | null;
    LeadRoleStatuses?: Array<LeadRoleStatusApi> | null;
};
// ...
export type SubStatusApi = {
    Id?: number;
    SubStatusName?: string | null;
    StatusScript?: string | null;
    Reasons?: string | null;
    IsClosed?: boolean;
    IsCurrentClient?: boolean;
    IsScheduleable?: boolean;
    IsIncludedInSearch?: boolean;
    IsValueEnabled?: boolean;
    IsOnDashboard?: boolean;
    IsResolved?: boolean;
    IsWanted?: boolean;
    ReferredType?: ReferredSubstatusType;
    DisplayOrder?: number;
};